Refactoring, fixes and improvements.

+++ src/main/java/diarsid/beam/core/modules/handlers/WebPagesHandlerWorker.java
@@ -6,9 +6,11 @@
 
 package diarsid.beam.core.modules.handlers;
 
+import java.util.Collections;
 import java.util.List;
 
 import diarsid.beam.core.entities.WebPage;
+import diarsid.beam.core.entities.WebPageDirectory;
 import diarsid.beam.core.entities.WebPagePlacement;
 import diarsid.beam.core.modules.IoInnerModule;
 import diarsid.beam.core.modules.data.DaoWebPages;
@@ -57,6 +59,9 @@ class WebPagesHandlerWorker implements WebPagesHandler {
     
     @Override
     public boolean deleteWebPage(String name, String dir, WebPagePlacement place) {        
+        if ( ! dir.matches(WEB_NAME_REGEXP) || ! name.matches(WEB_NAME_REGEXP)) {
+            return false;
+        }    
         name = name.trim().toLowerCase();
         return this.dao.deleteWebPage(name, dir, place);
     }
@@ -76,13 +81,29 @@ class WebPagesHandlerWorker implements WebPagesHandler {
             String directory, WebPagePlacement placement, boolean strict) {
         
         directory = directory.trim().toLowerCase();
+        if ( ! directory.matches(WEB_NAME_REGEXP) ) {
+            return Collections.emptyList();
+        }
         return this.dao.getAllWebPagesInDirectoryAndPlacement(
                 directory, placement, strict);
     }
     
     @Override
+    public List<WebPage> getWebPagesByNameInDirAndPlace(
+            String name, String dir, WebPagePlacement place) {
+        
+        if ( ! dir.matches(WEB_NAME_REGEXP) || ! name.matches(WEB_NAME_REGEXP)) {
+            return Collections.emptyList();
+        }
+        return this.dao.getWebPagesByNameInDirAndPlace(name, dir, place);
+    }
+    
+    @Override
     public List<WebPage> getWebPages(String name) {
         name = name.trim().toLowerCase();
+        if ( ! name.matches(WEB_NAME_REGEXP)) {
+            return Collections.emptyList();
+        }
         if (name.contains("-")){
             return this.dao.getWebPagesByNameParts(name.split("-"));
         } else {
@@ -128,11 +149,31 @@ class WebPagesHandlerWorker implements WebPagesHandler {
     @Override
     public boolean editWebPageOrder(
             String name, String dir, WebPagePlacement place, int newOrder) {
+        if ( newOrder <= 0 ) {
+            return false;
+        }
+        if ( ! dir.matches(WEB_NAME_REGEXP) || ! name.matches(WEB_NAME_REGEXP)) {
+            return false;
+        }
         return this.dao.editWebPageOrder(name, dir, place, newOrder);
     }
     
     @Override
-    public boolean renameDirectory(
+    public WebPageDirectory getDirectoryExact(WebPagePlacement place, String dir) {
+        dir = dir.trim().toLowerCase();
+        if ( ! dir.matches(WEB_NAME_REGEXP) ) {
+            return null;
+        } 
+        return this.dao.getDirectoryExact(place, dir);
+    }
+    
+    @Override
+    public List<WebPageDirectory> getAllDirectoriesIn(WebPagePlacement placement) {
+        return this.dao.getAllDirectoriesIn(placement);
+    }
+    
+    @Override
+    public boolean renameDirectoryInPlacement(
             String directory, String newDirectory, WebPagePlacement placement) {
         
         directory = directory.trim().toLowerCase();
@@ -140,24 +181,51 @@ class WebPagesHandlerWorker implements WebPagesHandler {
         if ( ! newDirectory.matches(WEB_NAME_REGEXP)) {
             return false;
         }
-        return this.dao.editDirectoryNameInPlacement(directory, newDirectory, placement);
+        return this.dao.renameDirectoryInPlacement(directory, newDirectory, placement);
     }
     
     @Override
     public boolean editDirectoryOrder(
             WebPagePlacement place, String name, int newOrder) {
         
+        if ( newOrder <= 0 ) {
+            return false;
+        }
+        if ( ! name.matches(WEB_NAME_REGEXP)) {
+            return false;
+        }
         return this.dao.editDirectoryOrder(place, name, newOrder);
     }
     
     @Override
+    public boolean deleteDirectoryAndPages(String dir, WebPagePlacement place) {
+        dir = dir.trim().toLowerCase();
+        if ( ! dir.matches(WEB_NAME_REGEXP) ) {
+            return false;
+        }        
+        WebPageDirectory dirToDelete = new WebPageDirectory(dir, place);
+        return this.dao.deleteDirectoryAndPages(dirToDelete);
+    }
+    
+    @Override
+    public boolean createEmptyDirectory(WebPagePlacement place, String dir) {
+        dir = dir.trim().toLowerCase();
+        if ( ! dir.matches(WEB_NAME_REGEXP) ) {
+            return false;
+        }
+        return this.dao.createEmptyDirectory(place, dir);
+    }
+    
+    @Override
     public boolean moveWebPageTo
-            (String pageName, String newDirectory, WebPagePlacement placement) {
-        
-        if ( ! newDirectory.matches(WEB_NAME_REGEXP)) {
+            (String pageName, String newDir, WebPagePlacement placement) {
+                
+        pageName = pageName.trim().toLowerCase();
+        newDir = newDir.trim().toLowerCase();
+        if ( ! newDir.matches(WEB_NAME_REGEXP) || ! pageName.matches(WEB_NAME_REGEXP)) {
             return false;
         }        
         return this.dao.moveWebPageToPlacementAndDirectory
-                (pageName, newDirectory, placement);
+                (pageName, newDir, placement);
     }
 }
